improve channel edit styling
