+++ src/ui/component/channelEdit/index.js
