+++ static/locales/en.json
@@ -460,5 +460,24 @@
   "LBRY names cannot contain that symbol ($, #, @)": "LBRY names cannot contain that symbol ($, #, @)",
   "Path copied.": "Path copied.",
   "Open Folder": "Open Folder",
-  "Create Backup": "Create Backup"
+  "Create Backup": "Create Backup",
+  "Submit": "Submit",
+  "Website": "Website",
+  "aprettygoodsite.com": "aprettygoodsite.com",
+  "yourstruly@example.com": "yourstruly@example.com",
+  "Thumbnail source": "Thumbnail source",
+  "Thumbnail (400x400)": "Thumbnail (400x400)",
+  "https://example.com/image.png": "https://example.com/image.png",
+  "Cover source": "Cover source",
+  "Cover (1000x300)": "Cover (1000x300)",
+  "Editing": "Editing",
+  "Edit Your Channel": "Edit Your Channel",
+  "Editing Your Channel": "Editing Your Channel",
+  "We can explain... We know this page won't win any design awards, we have a cool idea for channel edits in the future. We just wanted to release a very very very basic version that just barely kinda works so people can use it.": "We can explain... We know this page won't win any design awards, we have a cool idea for channel edits in the future. We just wanted to release a very very very basic version that just barely kinda works so people can use it.",
+  "We can explain... \n\n We know this page won't win any design awards, we have a cool idea for channel edits in the future. We just wanted to release a very very very basic version that just barely kinda works so people can use it.": "We can explain... \n\n We know this page won't win any design awards, we have a cool idea for channel edits in the future. We just wanted to release a very very very basic version that just barely kinda works so people can use it.",
+  "We can explain...": "We can explain...",
+  "We know this page won't win any design awards, we have a cool idea for channel edits in the future. We just wanted to release a very very very basic version that just barely kinda works so people can use": "We know this page won't win any design awards, we have a cool idea for channel edits in the future. We just wanted to release a very very very basic version that just barely kinda works so people can use",
+  "We know this page won't win any design awards, we just wanted to release a very very very basic version that just barely kinda works so people can use": "We know this page won't win any design awards, we just wanted to release a very very very basic version that just barely kinda works so people can use",
+  "We know this page won't win any design awards, we just wanted to release a very very very basic version that just barely kinda works so people can use it right now. There is a much nicer version in the works.": "We know this page won't win any design awards, we just wanted to release a very very very basic version that just barely kinda works so people can use it right now. There is a much nicer version in the works.",
+  "We know this page won't win any design awards, we just wanted to release a very very very basic version that just barely kinda works so people can use it right now. There is a much nicer version being worked on.": "We know this page won't win any design awards, we just wanted to release a very very very basic version that just barely kinda works so people can use it right now. There is a much nicer version being worked on."
 }
